Added number of satellites to SMSs.
We were reaching the 160 characters in some SMSs, so I decided to
cut the altitude and battery levels to integer precission. This
gives us enough precission to understand what is happening and
to make decisions, but does not fill the SMSs with small numbers.

For example, there is no need to know that the battery is at
94.400000%. If we get 94% is fine.

Moreover, some of these numbers only had one decimal point number,
so we were using two many characters for nothing.